// js/stat-evaluation-player/src/generated/BumpEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type BumpEvent = { time: number, frame: number, initiator: RemoteIdTs, victim: RemoteIdTs, initiator_is_team_0: boolean, victim_is_team_0: boolean, is_team_bump: boolean, strength: number, confidence: number, contact_distance: number, closing_speed: number, victim_impulse: number, initiator_position: [number, number, number], victim_position: [number, number, number], };