// js/stat-evaluation-player/src/generated/BumpPlayerStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BumpPlayerStats = { bumps_inflicted: number, bumps_taken: number, team_bumps_inflicted: number, team_bumps_taken: number, last_bump_time: number | null, last_bump_frame: number | null, last_bump_strength: number | null, max_bump_strength: number, cumulative_bump_strength: number, };